requirements.txt
flask
flask-cors
gunicorn
opencv-python-headless
numpy
firebase-admin
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
python-dotenv
opencv-python
scikit-learn
opencv-contrib-python